Refactor setPtsAndDuration API to use uint32_t duration parameter
Change both IsoBmffBuffer::setPtsAndDuration() and IsoBmffHelper::SetPtsAndDuration() to take uint32_t duration instead of uint64_t, reflecting the actual ISOBMFF spec constraint that sample duration fields are 32-bit.

Benefits:
- API signature now reflects actual constraints (no silent data corruption)
- Compile-time type safety - callers see the constraint explicitly
- Eliminates runtime validation and clamping code
- Simpler implementation with no casts needed internally
- Caller in streamabstraction.cpp now uses explicit uint32_t cast, making overflow handling visible

Also fixes type bug in streamabstraction.cpp where duration was cast to int64_t instead of uint64_t.

Updated all test mocks and fakes to match new signature.

VPAAMP-888

+++ isobmff/isobmffbuffer.h
@@ -285,9 +285,9 @@ public:
 	 *        duration will be set (if flagged as present).
 	 *
 	 * @param[in] pts - Base media decode time to set
-	 * @param[in] duration - Sample duration to set
+	 * @param[in] duration - Sample duration to set (uint32_t per ISOBMFF spec)
 	 */
-	void setPtsAndDuration(uint64_t pts, uint64_t duration);
+	void setPtsAndDuration(uint64_t pts, uint32_t duration);
 
 	/**
 	 * @fn getFirstPTS
